feat(filter)!: replace glob with regex
