+++ src/lib/cli-options.ts
@@ -38,7 +38,7 @@ export function getCliOptions() {
     )
     .option(
       "--filter <FILTER>",
-      "filter packages by wildcard pattern matching from package.json file",
+      "filter packages by regex pattern matching from package.json file",
     )
     .option(
       "--warn-days <NUMBER>",
@@ -62,8 +62,8 @@ export function getCliOptions() {
 Examples:
   $ npm-check-last-publish --sort name --order asc
   $ npm-check-last-publish --sort average
-  $ npm-check-last-publish --filter "@types/*"
-  $ npm-check-last-publish --filter "react-*"
+  $ npm-check-last-publish --filter @types
+  $ npm-check-last-publish --filter "^react.*s$"
   $ npm-check-last-publish --warn-days 60 --error-days 120
   $ npm-check-last-publish --output json > report.json
   $ npm-check-last-publish --output csv > report.csv
