+++ src/lib/expand-package-patterns.ts
@@ -1,14 +1,10 @@
-import { globRegex } from "./glob-regex.ts";
-
 export function expandPackagePatterns(
   pattern: string,
   allNames: string[],
 ): string[] {
-  if (!pattern.includes("*")) return [pattern];
-
-  const regex = globRegex(pattern);
   const matched = allNames.filter((pkg) => {
-    return regex.test(pkg);
+    return new RegExp(pattern).test(pkg);
   });
+
   return [...new Set(matched)];
 }
